fixed todoApp
