--- src/types/Loading.ts
@@ -1,6 +1,0 @@
-export type LoadingTypes = {
-  todos: boolean;
-  add: boolean;
-  deletedId: number[];
-  update: number | null;
-};
