--- src/types/Todo.ts
@@ -1,6 +1,0 @@
-export interface Todo {
-  id: number;
-  userId: number;
-  title: string;
-  completed: boolean;
-}
