--- src/types/consts.ts
@@ -1,1 +1,0 @@
-export const USER_ID = 3349;
